blueprint_gui/requirements.txt
PySide6>=6.5
pyyaml>=6.0
markdown>=3.5
